there were some simple late-night errors with JoyData::changes()

+++ exampl3/src/ofApp.cpp
@@ -5,7 +5,7 @@
 void ofApp::setup(){
 	ofBackground(22);
 	ofxGLFWJoystick::one().printJoystickList();
-    
+
 }
 
 //--------------------------------------------------------------
@@ -15,12 +15,14 @@ void ofApp::update(){
     
     auto changes = ofxGLFWJoystick::one().getChangedValues(0);
     
+    /*
     for(auto& kv : changes.buttons) {
-        ofLogNotice("exampl3") << "button " << kv.first << " = " << kv.second;
+        ofLogNotice("exampl3") << "button " << kv.first << " = " << static_cast<int>(kv.second);
     }
     for(auto& kv : changes.axes) {
         ofLogNotice("exampl3") << "axis " << kv.first << " = " << kv.second;
     }
+    */
 }
 
 //--------------------------------------------------------------
